+++ commandsManager.js
